ecampus/eCampusAPI/ecampus_api/requirements.txt
aniso8601==7.0.0
asgiref
boto3==1.18.48
botocore==1.21.65
certifi==2023.5.7
charset-normalizer
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-dbbackup==3.3.0
django-environ==0.10.0
django-filter==2.4.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-private-storage==2.2.2
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2
jmespath==0.10.0
MarkupSafe
mysqlclient
numpy
openpyxl==3.0.7
packaging
pandas
phonenumbers
Pillow
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
Rx==1.6.3
s3transfer==0.5.2
singledispatch
six==1.16.0
sqlparse==0.4.4
typing_extensions
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
zmq==0.0.0
